Moved BFHtppClient and ChannelServiceHandler out of skills directly into core
