+++ libraries/botbuilder-core/botbuilder/core/__init__.py
@@ -17,9 +17,11 @@ from .bot_state import BotState
 from .bot_state_set import BotStateSet
 from .bot_telemetry_client import BotTelemetryClient, Severity
 from .card_factory import CardFactory
+from .channel_service_handler import BotActionNotImplementedError, ChannelServiceHandler
 from .conversation_state import ConversationState
 from .intent_score import IntentScore
 from .invoke_response import InvokeResponse
+from .bot_framework_http_client import BotFrameworkHttpClient
 from .memory_storage import MemoryStorage
 from .memory_transcript_store import MemoryTranscriptStore
 from .message_factory import MessageFactory
@@ -44,6 +46,7 @@ __all__ = [
     "AnonymousReceiveMiddleware",
     "AutoSaveStateMiddleware",
     "Bot",
+    "BotActionNotImplementedError",
     "BotAdapter",
     "BotAssert",
     "BotFrameworkAdapter",
@@ -53,10 +56,12 @@ __all__ = [
     "BotTelemetryClient",
     "calculate_change_hash",
     "CardFactory",
+    "ChannelServiceHandler",
     "ConversationState",
     "conversation_reference_extension",
     "IntentScore",
     "InvokeResponse",
+    "BotFrameworkHttpClient",
     "MemoryStorage",
     "MemoryTranscriptStore",
     "MessageFactory",
