+++ samples/experimental/skills-prototypes/simple-bot-to-bot/simple-root-bot/bots/root_bot.py
@@ -2,11 +2,11 @@ from typing import List
 
 from botbuilder.core import (
     ActivityHandler,
+    BotFrameworkHttpClient,
     ConversationState,
     MessageFactory,
     TurnContext,
 )
-from botbuilder.core.integration import BotFrameworkHttpClient
 from botbuilder.core.skills import SkillConversationIdFactory
 
 from botbuilder.schema import ActivityTypes, ChannelAccount
